+++ test/unit/invariants.t.sol
@@ -133,7 +133,7 @@ contract InvariantSucker is JBSucker {
     }
 
     function test_setNumberOfClaimsSent(address token, uint256 count) external {
-        _outboxOf[token].numberOfClaimsSent = count;
+        _outboxOf[token].numberOfClaimsSent = uint192(count);
     }
 
     function test_setDeprecatedAfter(uint256 timestamp) external {
